requirements.txt
asgiref==3.4.0
astroid==2.5.6
attrs==20.3.0
autopep8==1.5.7
awscli==1.20.12
backcall==0.2.0
botocore==1.21.12
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
colorama==0.4.3
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.4
django-rest-auth==0.9.5
django-saml2-auth==2.2.1
django-saml2-auth-ai==2.1.6
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
docutils==0.15.2
elementpath==2.2.3
environs==9.3.2
Flask==2.0.0
Flask-Cors==3.0.10
idna==2.10
importlib-resources==5.2.0
iniconfig==1.1.1
ipdb==0.13.7
ipython==7.23.1
ipython-genutils==0.2.0
isort==5.8.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
marshmallow==3.12.1
matplotlib-inline==0.1.2
mccabe==0.6.1
packaging==20.9
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pip-autoremove==0.9.1
pluggy==0.13.1
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==1.7.1
pylint==2.8.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pysaml2==7.0.1
pytest==6.2.3
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
xmlschema==1.6.4
zipp==3.4.1
